+++ partitura/io/matchfile_utils.py
@@ -46,6 +46,7 @@ key_signature_pattern = re.compile(
 pitch_class_pattern = re.compile("(?P<step>[A-Ga-g])(?P<alter>[#bn]*)")
 
 number_pattern = re.compile(r"\d+")
+vnumber_pattern = re.compile(r"v\d+")
 
 # For matchfiles before 1.0.0.
 old_version_pattern = re.compile(r"^(?P<minor>[0-9]+)\.(?P<patch>[0-9]+)")
@@ -868,6 +869,35 @@ def format_time_signature(value: MatchTimeSignature) -> str:
 
 def format_time_signature_list(value: MatchTimeSignature) -> str:
     value.is_list = True
+    return str(value)
+
+
+class MatchTempoIndication(MatchParameter):
+    def __init__(
+        self,
+        value: str,
+        is_list: bool = False,
+    ):
+        super().__init__()
+        self.value = self.from_string(value)[0]
+        self.is_list = is_list
+
+    def __str__(self):
+        return self.value
+
+    @classmethod
+    def from_string(cls, string: str) -> MatchTempoIndication:
+        content = interpret_as_list(string)
+        return content
+
+
+def interpret_as_tempo_indication(value: str) -> MatchTempoIndication:
+    tempo_indication = MatchTempoIndication.from_string(value)
+    return tempo_indication
+
+
+def format_tempo_indication(value: MatchTempoIndication) -> str:
+    value.is_list = False
     return str(value)